+++ templates/user/base.html
@@ -31,11 +31,11 @@
 
               <!-- Link -->
               <li class="nav-item">
-                <a class="nav-link" href="/">Home</a>
+                <a class="nav-link" href= "{% url 'home' %}">Home</a>
               </li>
 
               <li class="nav-item">
-                <a class="nav-link" href="/signup">Signup</a>
+                <a class="nav-link" href="{% url 'user:signup' %}">Signup</a>
               </li>
 
               <li class="nav-item">
