+++ user/urls.py
@@ -4,6 +4,5 @@ from . import views
 app_name = 'user'
 
 urlpatterns = [
-    path('', views.HomeView.as_view(), name='home'),
     path('signup/', views.SignUpView.as_view(), name='signup'),
 ]
